# swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/photos/{user_id}": {
      "get": {
        "operationId": "PhotosController_getPhotoByUser",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "",
            "content": {
              "application/json": {
                "schema": {
                  "type": "string"
                }
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "PhotosController_deletePhoto",
        "parameters": [
          {
            "name": "user_id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "number"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          }
        }
      }
    },
...
  },
  "info": {
    "title": "ENVOCC CARD",
    "description": "Example.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
...
}